# reduce pointer count from 256 to 128